requirements.txt
# Core dependencies
sentence-transformers==2.7.0
huggingface_hub==0.20.0
transformers==4.36.0
torch>=1.13.0

# Vector database
chromadb==0.4.22

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Utilities
pyyaml>=6.0
tqdm>=4.60.0
Pillow>=9.0.0

# GUI (built-in with Python)
# tkinter is included with Python
